Create Calculator using Python
To make evaluations and manipulate the variables to make it work as per the need is what makes python programming an easy and significant tool. With the help of the below module, we shall be exploring the various python topics studied so far to apply the learnings and make a calculator program in python. In this article, We will learn a simple command-line calculator program in Python 3.